requirements.txt
pycrypto==2.6.1
inotify==0.2.9
